% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.blup.model.R
\name{BLUP}
\alias{BLUP}
\title{Calculate best linear unbiased predictors (BLUPs) for a mixed effects model.}
\usage{
BLUP(marker, measurement.time, fixed, random = NULL, id, data)
}
\arguments{
\item{marker}{name of continuous marker to be modeled using mixed effects model. This will be used as the outcome for the mixed effects model.}

\item{measurement.time}{name of time of measurement from baseline.}

\item{fixed}{name of variables to be used as fixed effects.}

\item{random}{name of variables to be used as random effects.}

\item{id}{character name of subject id in data}

\item{data}{data.frame consisting of id, marker, measurement.time, fixed and random effects used for modeling. Observations with missing data in any of these variables will be removed.}
}
\value{
An object of class "PC_BLUP" which is a list containing:

\item{model}{ A 'lme' fit object from the nlme package.}
\item{fit}{  A data.frame consisting of id, measurement time, marker, 'fitted.blup' giving the fitted blup estimates  for the marker at each measurement time for the complete data. There are also columns giving the fixed and random components of the blup for each variable in the mixed effect model. These could be used to calculate, for example, an blup estimate for the rate of change in marker values over time.
}
\item{id, marker, measurement.time, fixed, random}{Function inputs.}
}
\description{
Smooth predictors/markers through time using mixed effect models and estimate BLUPs.
}
\examples{


}
