% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PC.model.frame}
\alias{PC.model.frame}
\title{Create a model frame for a PC model fit.}
\usage{
PC.model.frame(object, newdata, ...)
}
\arguments{
\item{object}{object of class 'PC_cox' fit using the 'PC.Cox' function}

\item{newdata}{data.frame with new data for which to model frame components. All variables used to fit the PC.Cox model must be present. Observations with missing data will be removed.}
}
\value{
A data.frame copy of newdata with added columns showing BLUP estimates or spline bases for measurement time used to fit the PC cox model 'object'.
}
\description{
Extract the model frame for a PC model fit that employed natural splines for measurement times or marker smoothing via BLUP estimates.
}
\note{
if no BLUPs or natural splines are used in the fitting of PC model 'object', newdata will be returned with no addtional columns.
}
\examples{
data(pc_data)


#fit a model using natural cubic splines to model measurement time
# and BLUPs to smooth marker measurements.
pc.model.2 <-  PC.Cox(
 id = "sub.id",
 stime = "time",
 status = "status",
 measurement.time = "meas.time",
 markers = c("marker", "marker_2"),
 data = pc_data,
 use.BLUP = c(TRUE, TRUE),
 knots.measurement.time = 3)

#BLUPs are used in this pc model so marker trajectory is needed for
#BLUP estimation and risk prediction calculation
PC.model.frame(pc.model.2, newdata = newdata.subj.6)


}
