% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_data}
\alias{pc_data}
\title{Simulated dataset for package 'partlyconditional'}
\format{A data frame with 500 rows and 6 variables:
\itemize{
  \item Age: Age in years.
  \item Female: Indicator for female gender.
  \item Smokes: Indicator for smoking status.
  \item Marker1: simulated biomarker.
  \item Marker2: simulated biomarker.
  \item Cancer: Indicator for cancer.
}}
\usage{
data(pc_data)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(pc_data)
}
\keyword{datasets}
