% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.blup.model.R
\name{predict.PC_BLUP}
\alias{predict.PC_BLUP}
\title{Predict method for BLUP object.}
\usage{
\method{predict}{PC_BLUP}(x, newdata, ...)
}
\arguments{
\item{x}{object of class `PC_BLUP` output from the BLUP function}

\item{newdata}{data.frame of new observations for which to obtain fitted BLUP estimates.}

\item{\dots}{ignored.}
}
\value{
A data.frame consisting of id, measurement time, marker, 'fitted.blup' giving the fitted blup estimates for the marker at each measurement time for newdata. There are also columns giving the fixed and random components of the blup for each variable in the mixed effect model. These could be used to calculate, for example, an blup estimate for the rate of change in marker values over time.
}
\description{
Calculate best linear unbiased predictors (BLUPs) using a previously fit mixed effects model for new observations.
}
