% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PC.cox}
\alias{PC.cox}
\title{Fit a partly conditional Cox model}
\usage{
PC.cox(id, stime, status, measurement.time, markers, data,
  use.BLUP = rep(FALSE, length(markers)), knots.measurement.time = NULL)
}
\arguments{
\item{id}{name of numeric subject id in data}

\item{stime}{name of survival time, may be repeated across subj. id.}

\item{status}{name of survival status indicator, generally 1 = bad outcome/death, 0 alive/censored.}

\item{measurement.time}{name of time of measurement from baseline.}

\item{markers}{data.frame consisting of n x p markers with names.}

\item{data}{data.frame}

\item{use.BLUP}{a vector of logical variables, indicating for each marker whether best linear unbiased predictors (BLUPs) should be calculated. If true, a mixed effect model of the form `lme(marker ~ 1 + measurement.time, random = ~ 1 + measurement.time | id)`, will be fit univariately for each marker.}

\item{knots.measurement.time}{number of knots to use when modeling measurement.time using natural cubic splines (using function `ns`) in the partly conditional Cox model. Set to 'NA' if no splines are to be used.}
}
\value{
a 'PC_cox' model fit object, consisting of
}
\description{
later
}
\examples{


}
