% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_data}
\alias{pc_data}
\title{Simulated dataset for package 'partlyconditional'}
\format{A data frame with 478 observations from 100 subject
\itemize{
  \item sub.id: numeric subject id.
  \item time: survival time in months, repeated across id
  \item status: adverse outcome status
  \item meas.time: measurement time in months
  \item marker_1: simulated marker
  \item marker_2: simulated marker
}}
\usage{
data(pc_data)
}
\description{
Simulated dataset for package 'partlyconditional'
}
\examples{
data(pc_data)
}
\keyword{datasets}
