% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.PC_cox}
\alias{predict.PC_cox}
\title{Predict method for partly conditional cox models.}
\usage{
\method{predict}{PC_cox}(object, newdata, prediction.time, ...)
}
\arguments{
\item{object}{object of class 'PC.cox' fit using the 'PC.cox' function}

\item{prediction.time}{}
}
\value{

}
\description{
Estimate predicted absolute n-year risk from a PC.cox model object.
}
